pyproject.toml
[tool.poetry]
name = "importdata"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
arabic-reshaper = "^2.1.3"
openpyxl = "^3.0.7"
python-bidi = "^0.4.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
